clear di
